ARM: dts: am335x-bone-common: add cpu0 and mmc1 triggers
This matches the vendor 3.8.x configuration that is shipping
with the boards.

The LED layout is now:
 USR0: heartbeat
 USR1: mmc0 (micro-SD slot)
 USR2: cpu0
 USR3: mmc1 (eMMC)

The cpu0 triggers was put in between the mmc triggers to make
is easier to see where the disk activity is.

Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>
Tested-by: Kevin Hilman <khilman@linaro.org>
Reviewed-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Benoit Cousson <bcousson@baylibre.com>
(cherry picked from commit ec8a75979f199fd6fdcf24f8d1c928a48c61483e)